Move useAdminSetting hooks from basic/api to shared/common to fix boundary violation
The useAdminSetting and useAdminSettings hooks imported from shared/common
(useToast) which is not allowed from the basic/api tier. Relocated them to
metabase/common/hooks where the dependency is valid, and updated all 39
consumer imports across frontend and enterprise.
