+++ enterprise/frontend/src/metabase-enterprise/auth/containers/JwtAuthCard/JwtAuthCard.tsx
@@ -1,9 +1,8 @@
 import { t } from "ttag";
 
 import { AuthCard } from "metabase/admin/settings/auth/components/AuthCard";
-import { useAdminSetting } from "metabase/api/utils";
 import { LoadingAndErrorWrapper } from "metabase/common/components/LoadingAndErrorWrapper";
-import { useHasTokenFeature } from "metabase/common/hooks";
+import { useAdminSetting, useHasTokenFeature } from "metabase/common/hooks";
 import type { EnterpriseSettings } from "metabase-types/api";
 
 import { JWT_SCHEMA } from "../../constants";
