+++ enterprise/frontend/src/metabase-enterprise/auth/containers/OidcAuthCard/OidcAuthCard.tsx
@@ -2,10 +2,13 @@ import { useCallback } from "react";
 import { t } from "ttag";
 
 import { AuthCard } from "metabase/admin/settings/auth/components/AuthCard";
-import { useAdminSetting } from "metabase/api/utils";
 import { getErrorMessage } from "metabase/api/utils/errors";
 import { LoadingAndErrorWrapper } from "metabase/common/components/LoadingAndErrorWrapper";
-import { useHasTokenFeature, useToast } from "metabase/common/hooks";
+import {
+  useAdminSetting,
+  useHasTokenFeature,
+  useToast,
+} from "metabase/common/hooks";
 import {
   useDeleteCustomOidcMutation,
   useGetCustomOidcProvidersQuery,
