+++ enterprise/frontend/src/metabase-enterprise/semantic_search/SearchSettingsWidget.tsx
@@ -5,7 +5,8 @@ import _ from "underscore";
 import { SettingHeader } from "metabase/admin/settings/components/SettingHeader";
 import { BasicAdminSettingInput } from "metabase/admin/settings/components/widgets/AdminSettingInput";
 import { UpsellSemanticSearchPill } from "metabase/admin/upsells/UpsellSemanticSearch";
-import { getErrorMessage, useAdminSetting } from "metabase/api/utils";
+import { getErrorMessage } from "metabase/api/utils";
+import { useAdminSetting } from "metabase/common/hooks";
 import { getPlan, isProPlan } from "metabase/common/utils/plan";
 import type { SearchSettingsWidgetProps } from "metabase/plugins";
 import { getSetting } from "metabase/selectors/settings";
