+++ enterprise/frontend/src/metabase-enterprise/whitelabel/components/HelpLinkSettings/HelpLinkSettings.tsx
@@ -7,8 +7,9 @@ import {
   BasicAdminSettingInput,
   SetByEnvVar,
 } from "metabase/admin/settings/components/widgets/AdminSettingInput";
-import { getErrorMessage, useAdminSetting } from "metabase/api/utils";
+import { getErrorMessage } from "metabase/api/utils";
 import { ExternalLink } from "metabase/common/components/ExternalLink";
+import { useAdminSetting } from "metabase/common/hooks";
 import { Stack, TextInput } from "metabase/ui";
 import type { HelpLinkSetting } from "metabase-types/api";
