+++ frontend/src/metabase/admin/settings/auth/containers/GoogleAuthCard/GoogleAuthCard.tsx
@@ -1,7 +1,7 @@
 import { t } from "ttag";
 
-import { useAdminSetting } from "metabase/api/utils";
 import { LoadingAndErrorWrapper } from "metabase/common/components/LoadingAndErrorWrapper";
+import { useAdminSetting } from "metabase/common/hooks";
 import type { EnterpriseSettings } from "metabase-types/api";
 
 import { AuthCard } from "../../components/AuthCard";
