+++ frontend/src/metabase/admin/settings/components/EmbeddingSettings/EmbeddingToggle/EmbeddingToggle.tsx
@@ -2,8 +2,11 @@ import { useDisclosure } from "@mantine/hooks";
 import type { ChangeEvent } from "react";
 import { t } from "ttag";
 
-import { useAdminSetting, useAdminSettings } from "metabase/api/utils";
-import { useSetting } from "metabase/common/hooks";
+import {
+  useAdminSetting,
+  useAdminSettings,
+  useSetting,
+} from "metabase/common/hooks";
 import { Switch, type SwitchProps, Text } from "metabase/ui";
 
 import { EmbeddingLegaleseModal } from "../EmbeddingLegaleseModal";
