+++ frontend/src/metabase/admin/settings/components/UploadSettings/UploadSettingsForm.tsx
@@ -7,14 +7,14 @@ import {
   useListDatabasesQuery,
   useListSyncableDatabaseSchemasQuery,
 } from "metabase/api";
-import { getErrorMessage, useAdminSetting } from "metabase/api/utils";
+import { getErrorMessage } from "metabase/api/utils";
 import {
   ActionButton,
   type ActionButtonHandle,
 } from "metabase/common/components/ActionButton";
 import { Link } from "metabase/common/components/Link";
 import { LoadingAndErrorWrapper } from "metabase/common/components/LoadingAndErrorWrapper";
-import { useSetting, useToast } from "metabase/common/hooks";
+import { useAdminSetting, useSetting, useToast } from "metabase/common/hooks";
 import CS from "metabase/css/core/index.css";
 import {
   Alert,
