+++ frontend/src/metabase/admin/settings/components/widgets/HttpsOnlyWidget.tsx
@@ -1,8 +1,7 @@
 import { useCallback, useEffect, useState } from "react";
 import { t } from "ttag";
 
-import { useAdminSetting } from "metabase/api/utils";
-import { useHasTokenFeature } from "metabase/common/hooks";
+import { useAdminSetting, useHasTokenFeature } from "metabase/common/hooks";
 import { fetchWithTimeout } from "metabase/utils/fetchWithTimeout";
 
 import { AdminSettingInput } from "./AdminSettingInput";
