+++ frontend/src/metabase/api/utils/index.ts
@@ -1,3 +1,2 @@
 export * from "./errors";
-export * from "./settings";
 export * from "./use-token-refresh";
