+++ frontend/src/metabase/common/hooks/use-admin-setting/index.ts
@@ -1,0 +1,1 @@
+export { useAdminSetting, useAdminSettings } from "./use-admin-setting";
