+++ frontend/src/metabase/common/hooks/use-admin-setting/use-admin-setting.unit.spec.tsx
@@ -13,7 +13,7 @@ import {
   createMockSettings,
 } from "metabase-types/api/mocks";
 
-import { useAdminSetting, useAdminSettings } from "./settings";
+import { useAdminSetting, useAdminSettings } from "./use-admin-setting";
 
 const TestComponent = () => {
   const { value, settingDetails, updateSetting, isLoading } =
