+++ frontend/src/metabase/metabot/components/MetabotAdmin/MetabotConfig.tsx
@@ -13,11 +13,10 @@ import {
   useListMetabotsQuery,
   useUpdateMetabotMutation,
 } from "metabase/api";
-import { useAdminSetting } from "metabase/api/utils/settings";
 import { canonicalCollectionId } from "metabase/collections/utils";
 import { LoadingAndErrorWrapper } from "metabase/common/components/LoadingAndErrorWrapper";
 import { CollectionPickerModal } from "metabase/common/components/Pickers/CollectionPicker";
-import { useSetting, useToast } from "metabase/common/hooks";
+import { useAdminSetting, useSetting, useToast } from "metabase/common/hooks";
 import {
   FIXED_METABOT_ENTITY_IDS,
   FIXED_METABOT_IDS,
